chore: update node to 22.13.0
risk: nonprod
