+++ .nvmrc
@@ -1,1 +1,1 @@
-v18.17.0
+v22.13.0
