+++ common/config/rush/common-versions.json
@@ -21,7 +21,6 @@
         "@gooddata/frontend-npm-scripts": "1.2.0",
 
         "prettier": "~2.5.0",
-        "ts-node": "^10.4.0",
         "tslib": "^2.5.0",
         "typescript": "5.3.3",
