+++ rush.json
@@ -16,7 +16,7 @@
      * path segment in the "$schema" field for all your Rush config files.  This will ensure
      * correct error-underlining and tab-completion for editors such as VS Code.
      */
-    "rushVersion": "5.117.2",
+    "rushVersion": "5.148.0",
 
     /**
      * The next field selects which package manager should be installed and determines its version.
@@ -39,7 +39,7 @@
      * Specify a SemVer range to ensure developers use a Node.js version that is appropriate
      * for your repo.
      */
-    "nodeSupportedVersionRange": ">=18.17.0",
+    "nodeSupportedVersionRange": ">=22.13.0",
 
     /**
      * Odd-numbered major versions of Node.js are experimental.  Even-numbered releases
