+++ tools/app-toolkit/package.json
@@ -91,7 +91,6 @@
         "eslint-plugin-sonarjs": "^0.16.0",
         "eslint-plugin-tsdoc": "^0.2.14",
         "prettier": "~2.5.0",
-        "ts-node": "^10.4.0",
         "typescript": "5.3.3",
         "vitest": "1.0.4"
     }
