+++ tools/applink/package.json
@@ -70,7 +70,6 @@
         "json5": "^2.1.3",
         "lodash": "^4.17.21",
         "prettier": "~2.5.0",
-        "ts-node": "^10.4.0",
         "tslib": "^2.5.0",
         "typescript": "5.3.3",
         "vitest": "1.0.4"
