+++ tools/mock-handling/package.json
@@ -75,7 +75,6 @@
         "eslint-plugin-regexp": "^1.9.0",
         "eslint-plugin-sonarjs": "^0.16.0",
         "eslint-plugin-tsdoc": "^0.2.14",
-        "ts-node": "^10.4.0",
         "typescript": "5.3.3",
         "vitest": "1.0.4"
     }
